Add ability to browse through data by most recent (#105)
Adds a new sidebar panel named "feed" that shows issues ordered by id (highest to lowest, which approximately relates to reverse chronological order of submission). Clicking an issue in the sidebar feed zooms to and opens the issue on the map; clicking "open in sidebar" on a map pop-up will scroll to the issue 

A major addition with this PR is the addition of the Hash Router, which allows for the application state to be set by URL parameters (e.g. which issue is shown, which sidebar panel is selected). The behaviour and structure of the Hash Router is documented in the readme.