Add address search to parking map (#257)
Adds a search box to the /parking-map page that allows the user to search for a location by address or place name.

Address search is done using [photon](https://photon.komoot.io/). Search requests are tuned to the user's current map centre and limited to results within the bounding box for the City of Toronto. Typing in a search query will automatically start showing results; the user can then select one of the results to centre the map on that location. Search results can be cleared using the 'x' button.