Update parking data sources (#275)
Convert bicycle network and bicycle parking data resources to be configured via environment variable and switch both to use `data` branch of `parking-map-data` repo. 

This improves configurability and testability and allows for full separation of concerns between code and output data for the `parking-map-data` repo. The switch to using the `data` branch also facilitates data updates to be done daily instead of weekly.

Note that https://github.com/bikespace/parking-map-data/pull/13 should be merged after this to finalize the separation of production code and data in the [bikespace/parking-map-data](https://github.com/bikespace/parking-map-data) repo.