Parking map add filters (#273)
Adds two new UI elements to the parking map page: a filter control to show or hide features based on the values of one attribute key and a layer control to show or hide the bicycle network background layer. Both are organized within a <details> toggle box that is closed by default.

Additional notes and details are included in the PR for this commit.