Fix for flask-admin truncating microseconds
The wtforms widget that flask-admin uses doesn't include microseconds, which results in data loss or an error if the microsecond format is strictly enforced. This substitutes a custom version of the widget that preserves microseconds.

As a result, the custom version no longer shows the date picker and functions like a text input instead.